condaenv.3pi68g1n.requirements.txt
2captcha-python==1.2.1
addict==2.4.0
aiodns==3.0.0
aiohttp-retry==2.8.3
aleph-alpha-client==2.14.0
alephzero==0.3.17
annoy==1.17.2
anyio==3.7.1
appdirs==1.4.4
asttokens==2.2.1
attr-dicts==1.0.2
azure-cognitiveservices-speech==1.31.0
backcall==0.2.0
bardapi==0.1.32
bert-score==0.3.13
binpacking==1.5.2
blanc==0.3.1
blis==0.7.9
boto3==1.28.7
botocore==1.31.7
bottle==0.12.25
browser-cookie3==0.19.1
cachetools==5.3.1
catalogue==2.0.8
cattrs==22.2.0
cffi==1.15.1
cloudpickle==2.2.1
colorcet==3.0.1
comm==0.1.3
confection==0.1.0
configargparse==1.5.5
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
cython==3.0.0
dacite==1.6.0
datasets==2.5.2
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.4
dill==0.3.5.1
dm-control==1.0.13
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring-parser==0.15
dotmap==1.3.30
einops==0.6.1
emoji==2.6.0
evdev==1.6.1
executing==1.2.0
farama-notifications==0.0.4
flask==2.2.5
fonttools==4.40.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.4.0
gin-config==0.5.0
gitdb==4.0.10
gitpython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.64.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.59.1
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==20.1.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
h11==0.14.0
h2==4.1.0
hidapi==0.14.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.15.1
hyperframe==6.0.1
icetk==0.0.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-resources==5.10.4
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jax-jumpy==1.0.0
jedi==0.18.2
jeepney==0.8.0
jmespath==1.0.1
joblib==1.3.1
jsonlines==3.1.0
jsonpointer==2.3
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
jupyter-client==8.3.0
jupyter-core==5.3.1
jupyterlab-widgets==3.0.8
keyboard==0.13.5
kiwisolver==1.4.4
labmaze==1.0.6
langcodes==3.3.0
langdetect==1.0.9
lilac==0.1.1
llvmlite==0.39.1
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
mako==1.2.4
markdown==3.4.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mediapipe==0.10.2
moverscore==1.0.3
mujoco==2.3.6
multiprocess==0.70.13
murmurhash==1.0.9
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
numpy-quaternion==2022.4.3
oauthlib==3.2.2
open-clip-torch==2.20.0
open3d==0.17.0
openai==0.27.8
opencv-contrib-python==4.8.0.74
opencv-python==4.7.0.72
outcome==1.2.0
pandas==2.0.3
param==1.13.0
parameterized==0.9.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.0
portalocker==2.7.0
preshed==3.0.8
protobuf == 4.21.6
proto-plus==1.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.7.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pycryptodomex==3.18.0
pyct==0.5.0
pydantic==1.8.2
pydeprecate==0.3.2
pyemd==0.5.1
pyext==0.7
pygame==2.4.0
pyhocon==0.3.60
pymesh==1.0.2
pymongo==4.2.0
pymunk==6.5.1
pynput==1.7.6
pyparsing==3.0.9
pyquaternion==0.9.9
pyrealsense2==2.54.1.5217
python-xlib==0.33
pytorch-lightning==1.6.4
pytorch-pretrained-bert==0.6.2
pytrec-eval==0.5
pyzmq==25.1.0
referencing==0.29.1
regex==2023.6.3
requests-oauthlib==1.3.1
responses==0.18.0
retrying==1.3.4
revchatgpt==0.1.1
rich==13.4.2
rouge-score==0.1.2
rpds-py==0.8.10
rsa==4.9
s3transfer==0.6.1
sacrebleu==2.2.1
sacremoses==0.0.53
safetensors==0.3.1
scaleapi==2.13.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.11.2
segment-anything==1.0
selenium==4.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.26.0
setproctitle==1.3.2
sharedmem==0.3.8
simple-slurm==0.2.6
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
spacy==3.2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlitedict==1.7.0
srsly==2.4.7
stack-data==0.6.2
stanza==1.5.0
summ-eval==0.892
surge-api==1.1.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
thinc==8.0.17
threadpoolctl==3.1.0
tiktoken==0.3.3
timm==0.9.2
tls-client==0.2.1
tokenizers==0.13.3
torch==1.12.1
torchmetrics==0.11.4
torchvision==0.13.1
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
trio==0.22.2
trio-websocket==0.10.3
typed-argument-parser==1.8.0
typer==0.4.2
typing==3.7.4.3
typing-extensions==4.5.0
typing-inspect==0.9.0
uncertainty-calibration==0.1.4
undetected-chromedriver==3.5.0
uritemplate==4.1.1
urllib3==1.26.16
wandb==0.15.4
wasabi==0.10.1
websocket-client==1.3.3
websockets==11.0.3
werkzeug==2.2.3
widgetsnbextension==4.0.8
wsproto==1.2.0
xlrd==2.0.1
xxhash==3.2.0
yacs==0.1.8
yapf==0.40.1
zstandard==0.18.0